Extract finding extraction, author tier, specialist logs, and mention write formatters.
Move inline comment extraction, author tier resolution, shadow specialist publication logging, and mention PR/commit formatting out of the handler god-modules; drop duplicated dead author-tier helpers from review.ts.
