+++ src/handlers/mention.ts
@@ -56,6 +56,13 @@ import {
   isDiffSeekingMentionRequest,
   isReviewRequest,
 } from "./mention-request-classification.ts";
+import {
+  generateCommitSubject,
+  generatePrBody,
+  generatePrTitle,
+  parseWriteIntent,
+  summarizeWriteRequest,
+} from "./mention-write-formatters.ts";
 import { buildIssueCodeContext } from "../execution/issue-code-context.ts";
 import { buildMentionPrompt, buildMentionPromptDetails } from "../execution/mention-prompt.ts";
 import { buildReviewPrompt, buildReviewPromptDetails, matchPathInstructions } from "../execution/review-prompt.ts";
@@ -443,196 +450,6 @@ export function createMentionHandler(deps: {
         prConversationTouchedAt.delete(k);
       }
     }
-  }
-
-  function parseWriteIntent(userQuestion: string): {
-    writeIntent: boolean;
-    keyword: "apply" | "change" | "plan" | undefined;
-    request: string;
-  } {
-    const trimmed = userQuestion.trimStart();
-    const lower = trimmed.toLowerCase();
-
-    for (const keyword of ["apply", "change", "plan"] as const) {
-      const prefix = `${keyword}:`;
-      if (lower.startsWith(prefix)) {
-        return {
-          writeIntent: true,
-          keyword,
-          request: trimmed.slice(prefix.length).trim(),
-        };
-      }
-    }
-
-    return { writeIntent: false, keyword: undefined, request: userQuestion.trim() };
-  }
-
-
-  function summarizeWriteRequest(request: string): string {
-    const condensed = request
-      .replace(/\s+/g, " ")
-      .replace(/^[@`'"([{\s]+/, "")
-      .replace(/[@`'"\])}\s]+$/, "")
-      .replace(/^(?:can|could|would|will)\s+you\s+/i, "")
-      .replace(/^(?:please\s+)+/i, "")
-      .replace(/[?.!]+$/, "")
-      .trim();
-
-    const fallback = "requested update";
-    const normalized = condensed.length > 0 ? condensed : fallback;
-    const maxLen = 72;
-    return normalized.length <= maxLen ? normalized : `${normalized.slice(0, maxLen - 3).trimEnd()}...`;
-  }
-
-  function generatePrTitle(issueTitle: string | null, requestSummary: string, isFromPr: boolean): string {
-    const maxLen = 72;
-
-    if (issueTitle && issueTitle.trim().length > 0) {
-      // Clean the issue title: remove leading [tags], trailing issue refs
-      const cleaned = issueTitle
-        .replace(/^\[.*?\]\s*/g, "")
-        .replace(/\s*#\d+\s*$/, "")
-        .trim();
-
-      // Detect prefix from content
-      const lower = cleaned.toLowerCase();
-      let prefix: string;
-      if (/\b(?:fix|bug|crash|broken|error)\b/.test(lower)) {
-        prefix = "fix";
-      } else if (/\brefactor\b/.test(lower)) {
-        prefix = "refactor";
-      } else if (/\b(?:add|support|implement|feature|new)\b/.test(lower)) {
-        prefix = "feat";
-      } else {
-        prefix = isFromPr ? "fix" : "feat";
-      }
-
-      const full = `${prefix}: ${cleaned}`;
-      return full.length <= maxLen ? full : `${full.slice(0, maxLen - 3).trimEnd()}...`;
-    }
-
-    // Fallback: no issue title available
-    const defaultPrefix = isFromPr ? "fix" : "feat";
-    const full = `${defaultPrefix}: ${requestSummary}`;
-    return full.length <= maxLen ? full : `${full.slice(0, maxLen - 3).trimEnd()}...`;
-  }
-
-  function generateCommitSubject(params: {
-    issueTitle: string | null | undefined;
-    requestSummary: string;
-    isFromPr: boolean;
-    ref?: string; // e.g. "#27954" or "PR #42"
-  }): string {
-    const maxLen = 72;
-    const { issueTitle, requestSummary, isFromPr, ref } = params;
-
-    let subject: string;
-
-    if (issueTitle && issueTitle.trim().length > 0) {
-      const cleaned = issueTitle
-        .replace(/^\[.*?\]\s*/g, "")
-        .replace(/\s*#\d+\s*$/, "")
-        .trim();
-
-      const lower = cleaned.toLowerCase();
-      let prefix: string;
-      if (/\b(?:fix|bug|crash|broken|error)\b/.test(lower)) {
-        prefix = "fix";
-      } else if (/\brefactor\b/.test(lower)) {
-        prefix = "refactor";
-      } else if (/\b(?:add|support|implement|feature|new)\b/.test(lower)) {
-        prefix = "feat";
-      } else {
-        prefix = isFromPr ? "fix" : "feat";
-      }
-      subject = `${prefix}: ${cleaned}`;
-    } else {
-      const defaultPrefix = isFromPr ? "fix" : "feat";
-      subject = `${defaultPrefix}: ${requestSummary}`;
-    }
-
-    // Append ref if provided
-    if (ref) {
-      const withRef = `${subject} (${ref})`;
-      if (withRef.length <= maxLen) {
-        subject = withRef;
-      }
-      // If adding ref would exceed maxLen, truncate subject part to fit
-      else {
-        const refSuffix = ` (${ref})`;
-        const available = maxLen - refSuffix.length - 3; // 3 for "..."
-        if (available > 10) {
-          subject = `${subject.slice(0, available).trimEnd()}...${refSuffix}`;
-        }
-        // else just truncate without ref
-      }
-    }
-
-    return subject.length <= maxLen ? subject : `${subject.slice(0, maxLen - 3).trimEnd()}...`;
-  }
-
-  function generatePrBody(params: {
-    summary: string;
-    issueTitle: string | null;
-    sourceUrl: string;
-    triggerCommentUrl: string;
-    deliveryId: string;
-    headSha: string;
-    isFromPr: boolean;
-    issueNumber: number;
-    prNumber: number | undefined;
-    diffStat: string;
-    warnings?: string[];
-  }): string {
-    const {
-      summary, issueTitle, sourceUrl, triggerCommentUrl,
-      deliveryId, headSha, isFromPr, issueNumber, prNumber, diffStat,
-    } = params;
-
-    // Summary paragraph: prefer issue title context, fall back to request summary
-    const summaryParagraph = issueTitle && issueTitle.trim().length > 0
-      ? issueTitle.trim()
-      : summary;
-
-    const resolveOrRelate = isFromPr
-      ? `Related to #${prNumber}`
-      : `Resolves #${issueNumber}`;
-
-    const lines: string[] = [
-      summaryParagraph,
-      "",
-    ];
-
-    if (diffStat) {
-      lines.push("## Changes", "", diffStat, "");
-    }
-
-    if (params.warnings && params.warnings.length > 0) {
-      lines.push(
-        "## Automated warnings",
-        "",
-        ...params.warnings.map((w) => `- ${w}`),
-        "",
-      );
-    }
-
-    lines.push(
-      "---",
-      "",
-      resolveOrRelate,
-      "",
-      "<details>",
-      "<summary>Metadata</summary>",
-      "",
-      `- Source: ${sourceUrl}`,
-      `- Trigger: ${triggerCommentUrl}`,
-      `- Delivery: ${deliveryId}`,
-      `- Commit: ${headSha}`,
-      "",
-      "</details>",
-    );
-
-    return lines.join("\n");
   }
 
   async function scanDiffForFabricatedContent(dir: string): Promise<string[]> {
